+++ Sources/TuistGenerator/Generator/Generator.swift
@@ -88,7 +88,6 @@ public class Generator: Generating {
 
     public func generateProject(at path: AbsolutePath) throws -> AbsolutePath {
         let (graph, project) = try graphLoader.loadProject(path: path)
-
         let generatedProject = try projectGenerator.generate(project: project,
                                                              graph: graph,
                                                              sourceRootPath: path)
@@ -100,7 +99,7 @@ public class Generator: Generating {
         let tuistConfig = try graphLoader.loadTuistConfig(path: path)
         let (graph, project) = try graphLoader.loadProject(path: path)
 
-        let workspace = Workspace(name: project.name,
+        let workspace = Workspace(name: project.fileName,
                                   projects: graph.projectPaths,
                                   additionalFiles: workspaceFiles.map(FileElement.file))
 
@@ -114,7 +113,6 @@ public class Generator: Generating {
                                   workspaceFiles: [AbsolutePath]) throws -> AbsolutePath {
         let (graph, workspace) = try graphLoader.loadWorkspace(path: path)
         let tuistConfig = try graphLoader.loadTuistConfig(path: path)
-
         let updatedWorkspace = workspace
             .merging(projects: graph.projectPaths)
             .adding(files: workspaceFiles)