+++ Tests/TuistGeneratorTests/Models/TestData/Project+TestData.swift
@@ -5,6 +5,7 @@ import Foundation
 extension Project {
     static func test(path: AbsolutePath = AbsolutePath("/test/"),
                      name: String = "Project",
+                     fileName: String? = nil,
                      settings: Settings = Settings.test(),
                      filesGroup: ProjectGroup = .group(name: "Project"),
                      targets: [Target] = [Target.test()],
@@ -12,6 +13,7 @@ extension Project {
                      additionalFiles: [FileElement] = []) -> Project {
         return Project(path: path,
                        name: name,
+                       fileName: fileName,
                        settings: settings,
                        filesGroup: filesGroup,
                        targets: targets,