+++ fixtures/ios_app_with_frameworks/TuistConfig.swift
@@ -2,6 +2,7 @@ import ProjectDescription
 
 let config = TuistConfig(
   generationOptions: [
-    .generateManifest
+    .generateManifest,
+    .xcodeProjectName("AwesomePrefix-\(.projectName)-AwesomeSuffix")
   ]
 )